modified:   app/Http/Controllers/CheckoutController.php 	new file:   app/Http/Controllers/PaymentController.php 	modified:   app/Http/Controllers/PlaceOrderController.php 	modified:   app/Models/Order.php 	new file:   app/Models/OrderItem.php 	modified:   config/services.php 	modified:   database/migrations/2025_09_16_154721_add_profile_photo_path_to_users_table.php 	new file:   database/migrations/2025_09_22_074009_create_orders_table.php 	new file:   database/migrations/2025_09_22_074012_create_order_items_table.php 	modified:   resources/views/checkout.blade.php 	modified:   resources/views/home.blade.php 	modified:   resources/views/layout.blade.php 	modified:   resources/views/place-order.blade.php 	modified:   routes/web.php

+++ database/migrations/2025_09_16_154721_add_profile_photo_path_to_users_table.php
@@ -12,8 +12,11 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            // 'email' column pachi profile picture no path save karva mate navi column.
-            $table->string('profile_photo_path', 2048)->nullable()->after('email');
+            // Check if the column does not exist before adding it
+            if (!Schema::hasColumn('users', 'profile_photo_path')) {
+                // 'email' column pachi profile picture no path save karva mate navi column.
+                $table->string('profile_photo_path', 2048)->nullable()->after('email');
+            }
         });
     }
 
@@ -23,8 +26,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            // Jo migration ne revert karvi hoy to column ne kadhi nakhvani.
-            $table->dropColumn('profile_photo_path');
+            // Check if the column exists before dropping it
+            if (Schema::hasColumn('users', 'profile_photo_path')) {
+                $table->dropColumn('profile_photo_path');
+            }
         });
     }
 };
